+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## WIP
+## 3.5.16 (December 11, 2020)
 
 ### Fixed